requirements.txt
python-binance
TA-Lib
numpy
backtrader
flask

PS C:\Users\orotm\OneDrive\Asztali gép\Binance API> c:\users\orotm\appdata\local\packages\pythonsoftwarefoundation.python.3.8_qbz5n2kfra8p0\localcache\local-packages\python38\scripts\flask.exe run   